:sparkles: Add sparklines library
